+++ drf_scaffold/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'core'
+    'drf_scaffold_core'
 ]
 
 MIDDLEWARE = [
