+++ drf_scaffold_core/scaffold_templates/admin_templates.py
