+++ drf_scaffold_core/scaffold_templates/model_templates.py
